requirements.txt
numpy
pandas
streamlit
scikit-learn
pillow
pydantic